Arrows properly showing
I was unable to calculate the size of a rectangle that would contain the full arrow. So I ended opting for a "worst-case scenario" situation where the surface is a bit bigger than it is necessary. The workaround was to add an offset to each margin of the surface that would contain the arrow in such scenario (for example a full horizontal arrow where the full offset is used).

Also, added functionality to allow to delete all bodies by pressing space.