Added basic cinematic support
The velocity gets properly set depending on the arrow length and the objects move in uniform rectilinear motion accordingly to it.

If an object gets too far away (given by PLANET_MAX_DISTANCE constant), it gets destroyed. This will come in handy in the future when objects could get in hyperbolic trajectories and never come back.