Added arrows showing vel and acc during simulation
Changes in the arrow initializer that now accepts all parameters used so it can show different kinds of arrows (used now to show the simulation arrows with different colours and thickness).

Fixed the force not reseting to 0 after each frame.

When selecting velocity you can now right click to set it to zero.